refactor : applied the VAC pattern to TodoModal
