+++ src/screens/Home.tsx
@@ -8,7 +8,7 @@ import * as todoSlice from "../store/slices/todos"
 import * as modalSlice from "../store/slices/modal"
 
 import Todo from "../components/todo/Todo"
-import TodoModal from "../components/modal/TodoModal"
+import TodoModal from "../components/modal/TodoModal/TodoModal"
 import ControlBottomBar from "../components/ControlBottomBar"
 
 type HomeProps = {
